refactor: Page 관련 Controller와 API Controller 분리
- api 관련 경로 수정
