+++ src/main/java/shopping/ui/CartPageController.java
@@ -1,12 +1,12 @@
-package shopping.controller;
+package shopping.ui;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
-public class CartController {
+public class CartPageController {
 
-    @GetMapping("/cart")
+    @GetMapping("/cart-items")
     public String renderCartPage() {
         return "cart";
     }
