+++ src/main/java/shopping/ui/LoginPageController.java
@@ -1,0 +1,13 @@
+package shopping.ui;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class LoginPageController {
+
+    @GetMapping("/login")
+    public String renderLoginPage() {
+        return "/login";
+    }
+}
