+++ src/main/java/shopping/ui/OrderPageController.java
@@ -1,4 +1,4 @@
-package shopping.controller;
+package shopping.ui;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
