+++ src/main/resources/templates/index.html
@@ -12,7 +12,7 @@
   <nav class="nav flex justify-end">
     <div class="flex items-center gap-15">
       <a href="/" class="nav-button">상품</a>
-      <a href="/cart" class="nav-button">장바구니</a>
+      <a href="/cart-items" class="nav-button">장바구니</a>
       <a href="/orders" class="nav-button">주문목록</a>
       <a href="/login" class="nav-button">로그인</a>
     </div>
